requirements.txt
discord
PyYAML